Moved nodes.go to the node package
